+++ package.json
@@ -4,9 +4,14 @@
     "packages/*"
   ],
   "scripts": {
-    "test": "node --experimental-vm-modules $(yarn bin jest)"
+    "test": "node --experimental-vm-modules ./node_modules/.bin/jest",
+    "start": "concurrently --names=format,web,jest \"cd ./packages/format && yarn watch\" \"cd ./packages/web && yarn start\" \"yarn test --watch\"",
+    "postinstall": "lerna run prepare"
   },
   "devDependencies": {
-    "lerna": "^8.0.2"
+    "concurrently": "^8.2.2",
+    "jest": "^29.7.0",
+    "lerna": "^8.0.2",
+    "nodemon": "^3.0.2"
   }
 }
