+++ packages/tests/package.json
@@ -10,12 +10,11 @@
     "test": "node --experimental-vm-modules $(yarn bin jest)"
   },
   "devDependencies": {
+    "@ethdebug/format": "^0.1.0-0",
     "@hyperjump/json-schema": "1.6.7",
     "@jest/globals": "^29.7.0",
     "jest": "^29.7.0",
-    "json-schema-typed": "^8.0.1",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.3.3",
-    "yaml": "^2.3.4"
+    "typescript": "^5.3.3"
   }
 }
