+++ packages/web/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@docusaurus/core": "^3.0.1",
     "@docusaurus/preset-classic": "^3.0.1",
+    "@ethdebug/format": "^0.1.0-0",
     "@fortawesome/fontawesome-svg-core": "^6.5.1",
     "@fortawesome/free-brands-svg-icons": "^6.5.1",
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
@@ -24,19 +25,16 @@
     "@mdx-js/react": "^3.0.0",
     "clsx": "^1.2.1",
     "docusaurus-json-schema-plugin": "^1.10.0",
-    "jsonpointer": "^5.0.1",
     "prism-react-renderer": "^2.1.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-markdown": "^9.0.1",
-    "yaml": "^2.3.4",
     "yaml-template": "^1.0.0"
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "^3.0.1",
     "@docusaurus/tsconfig": "^3.0.1",
     "@docusaurus/types": "^3.0.1",
-    "raw-loader": "^4.0.2",
     "typescript": "~5.2.2"
   },
   "browserslist": {
