+++ packages/web/spec/type/overview.mdx
@@ -4,7 +4,7 @@ sidebar_position: 1
 
 import Tabs from "@theme/Tabs";
 import TabItem from "@theme/TabItem";
-import { describeSchema } from "@site/src/schemas";
+import { describeSchema } from "@ethdebug/format";
 import CodeBlock from "@theme/CodeBlock";
 
 # Overview
