--- packages/web/types.d.ts
@@ -1,4 +1,0 @@
-declare module "*.yaml" {
-  const content: string;
-  export default content;
-}
